Fix jitter calculation (#1019)

+++ worker/src/RTC/RtpStreamRecv.cpp
@@ -645,8 +645,9 @@ namespace RTC
 		if (this->params.clockRate == 0u)
 			return;
 
-		auto transit =
-		  static_cast<int>(DepLibUV::GetTimeMs() - (rtpTimestamp * 1000 / this->params.clockRate));
+		// NOTE: Based on https://github.com/versatica/mediasoup/issues/1018.
+		auto transit = static_cast<int>(
+		  DepLibUV::GetTimeMs() - (static_cast<uint64_t>(rtpTimestamp) * 1000 / this->params.clockRate));
 		int d = transit - this->transit;
 
 		// First transit calculation, save and return.
